Added Linear Slide
slide motor should be called "slide_motor" in the config

gamepad 2 can raise or lower the linear slide with the "y" and "a"
buttons.  There are limits placed on the maximum lift of the linear
slide, to avoid damage due to running it too far.  Currently this limit
is extremely small, so the linear slide has minimal movement available
to it.

also changed the 0.25 speed to activate sooner when the trigger is
depressed